-Added my own assert macro and replaced existing random error handling and asserts to just use this for general logic errors. -Trying to clean up a bit of the api for render queue and Command Builder.  Still need to make commandbuilder just a set of methods with no state. -Added the beginnings of a general texture utility -fixed a random slowdown that occurred when I switched the renderqueue to use the heap.  Uninitialized variables are bad. -

+++ RenderMode/Types/ModelManager.cpp
@@ -85,7 +85,7 @@ Model * ModelManager::loadModelGeometry(std::string &fname, RenderContext &rende
     int numIndices = 0;
     for (const auto &g : obj.groups) {
         for (const auto &f : g.second.faces) {
-            numIndices += f.size();
+            numIndices += (int)f.size();
         }
     }
     // TODO should this handle when we only have some of the components or is v/t/n always going to be specified?
